Bootloader

+++ Custom Bootloader STM32F103/Bootloader/Inc/MCAL/FLASH/flash_interface.h
@@ -29,7 +29,7 @@
 
 void 		 MFLASH_lock (void);
 void 		 MFLASH_unlock (void);
-ReturnStatus MFLASH_programPage (const u32  desiredAddress,const u8 desiredValue);
+ReturnStatus MFLASH_programPage (const u32  desiredAddress,const u16 desiredValue);
 ReturnStatus MFLASH_erasePage (const u32  desiredAddress);
 ReturnStatus MFLASH_massErase(void);
